operator/requirements.txt
aiohappyeyeballs==2.3.5
aiohttp==3.10.1
aiosignal==1.3.1
asn1crypto==1.5.1
attrs==24.1.0
cachetools==5.4.0
certbuilder==0.14.2
certifi==2024.7.4
certvalidator==0.11.1
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
frozenlist==1.4.1
google-auth==2.32.0
idna==3.7
iso8601==2.1.0
kopf==1.37.2
kubernetes==30.1.0
multidict==6.0.5
oauthlib==3.2.2
oscrypto==1.3.0
pyasn1==0.6.0
pyasn1_modules==0.4.0
pykube-ng==23.6.0
pyngrok==7.2.0
python-dateutil==2.9.0.post0
python-json-logger==2.0.7
PyYAML==6.0.2
requests==2.32.3
requests-oauthlib==2.0.0
rsa==4.9
six==1.16.0
typing_extensions==4.12.2
urllib3==2.2.2
websocket-client==1.8.0
yarl==1.9.4
